Detail and Reference activities moved to new package
