+++ app/src/main/res/layout/activity_place_detail.xml
@@ -6,7 +6,7 @@
 	xmlns:tools="http://schemas.android.com/tools"
 	android:layout_width="match_parent"
 	android:layout_height="match_parent"
-	tools:context="com.sygic.travel.sdkdemo.PlaceDetailActivity">
+	tools:context="com.sygic.travel.sdkdemo.detail.PlaceDetailActivity">
 
 	<android.support.constraint.ConstraintLayout
 		android:layout_width="match_parent"
